requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
python-dotenv
pgvector-python
pgvector
numpy sqlalchemy
scipy
faiss-cpu
faiss-gpu